feat: Register keyboard controller in Pheripherals class
This commit adds the functionality to register the "keyboard" controller in the `Pheripherals` class. The `register_controller` method is implemented to import the corresponding module and append it to the `controllers` list. This addition enhances the `Pheripherals` class by enabling support for the keyboard controller, improving the overall functionality of the PiPerW application.