feat: Add default theme with frame support
This commit adds a new file `default.py` under the `PiPerW/themes/default` directory. The `Theme` class is implemented as the default theme for the PiPerW application. It inherits from the `_base_theme_.BaseTheme` class and provides support for frames. The `frameArray` attribute is populated with frames from the `frames/default` folder. This addition enhances the functionality of the PiPerW application by introducing a default theme with frame support.